refactor: clear CodeScene findings on the token path
Excess Number of Function Arguments (max 4) flagged five functions in
token.go and one in token_lease.go. All of them threaded the same bundle,
so this fixes the shape once rather than each function in turn:

- token.go: mintOrAdopt, waitForLeaseOrAdoption, pollLeaseOrAdoption,
  mintLeased and persistAsyncThen carried some subset of (creds, source,
  io, lease), all fixed for a Source's lifetime with only ctx and forbid
  varying per call. They are now methods on a minter that owns the four,
  built once in NewStoredUserTokenSource.
- token_lease.go: acquireValkeyLock/releaseValkeyLock become methods on a
  valkeyLock owning the client, key and owner, built once in newMintLease.

token_test.go tripped Low Cohesion at 21 functions across 10
responsibilities against a threshold of 3. Split by seam into adoption,
mint-lease coordination, invalidation, refresh mechanics and shared
fakes. The in-flight-refresh test's setup moved into helpers; every
assertion it makes is unchanged.

broadcaster_test.go had two structurally identical sweep tests, now one
table-driven test covering both cases.

The tests that were verified to fail without their fix still do:
removing storeIfGen's generation check fails the in-flight-refresh test,
and dropping mintLeaseTTL below MaxMintLeaseHold() fails the lease
invariant test, which still computes both from the live constants.

hydrator.go's duplication finding is left alone deliberately. The three
fill functions differ in seed handling and each builds a distinct reply
type; a generic extraction would trade readable duplication for
indirection, and the file already improved from 8.94 to 9.10.

No suppressions.
